feat: Add battery voltage compensation
As battery voltage decreases, motors will run slower.
For the launcher, this results in inconsistent
scoring as the battery depletes. To compensate for
this, a new BatteryVoltageCompensator is implemented
to adjsut the motor power accordingly based on
battery voltage to ensure the launcher (and intake)
motors run at roughly the same speed, even as battery
depletes.